validating arguments

+++ ambilight/ambilight.go
@@ -24,12 +24,11 @@ type Ambilight struct {
 }
 
 // AmbilightConfiguration holds the configuration for the ambilight.
-// TODO: handle negative values
 type AmbilightConfiguration struct {
 	// target frames per second
-	Fps int
+	Fps uint
 	// max number of worker threads
-	Workers int
+	Workers uint
 }
 
 // Go fires up the ambilight.
@@ -41,7 +40,7 @@ func (a *Ambilight) Go() error {
 	frameDurationTarget := time.Duration(1000/a.Config.Fps) * time.Millisecond
 
 	// Create worker pool that can perform as many tasks as possible, given the allowed workers
-	maxParallelTasks := max(len(a.Screen.Areas), len(a.Universes))
+	maxParallelTasks := max(uint(len(a.Screen.Areas)), uint(len(a.Universes)))
 	workerPool := newWorkerPool(min(a.Config.Workers, maxParallelTasks))
 
 	for {
@@ -89,7 +88,6 @@ func (a *Ambilight) Go() error {
 }
 
 func getColorJob(a *Ambilight, area capture.Area) func() {
-	// TODO: maybe link directly in Area?
 	devices, ok := a.Mappings[area.ImageData.Borders]
 	if !ok {
 		// This area has no devices mapped.
@@ -111,14 +109,14 @@ func getColorJob(a *Ambilight, area capture.Area) func() {
 
 }
 
-func min(a int, b int) int {
+func min(a uint, b uint) uint {
 	if a < b {
 		return a
 	}
 	return b
 }
 
-func max(a int, b int) int {
+func max(a uint, b uint) uint {
 	if a > b {
 		return a
 	}
@@ -126,5 +124,4 @@ func max(a int, b int) int {
 }
 
 // Mapping holds a mapping from screen areas to DMX devices.
-// FIXME: it makes no sense to have more than one device per area as this would overwrite colors
 type Mapping map[*image.Rectangle][]*dmx.Device
